Separated UserNotifier to 3 notifiers, ProxyProvider providing filtered users
