+++ lib/main.dart
@@ -1,4 +1,6 @@
-import 'package:first_project/providers/user_notifier.dart';
+import 'package:first_project/notifiers/all_users_notifier.dart';
+import 'package:first_project/notifiers/filters_notifier.dart';
+import 'package:first_project/notifiers/search_query_notifier.dart';
 import 'package:first_project/routes/app_routes.dart';
 import 'package:first_project/screens/home_page.dart';
 import 'package:flutter/material.dart';
@@ -13,8 +15,17 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return ChangeNotifierProvider(
-      create: (_) => UserNotifier(),
+    return MultiProvider(
+      providers: [
+        ChangeNotifierProvider(create: (_) => FiltersNotifier()),
+        ChangeNotifierProvider(create: (_) => SearchQueryNotifier()),
+        ChangeNotifierProxyProvider2(
+          create: (BuildContext context) => AllUsersNotifier(context.read<FiltersNotifier>(), context.read<SearchQueryNotifier>())..fetchUsers(),
+          update: (BuildContext context, FiltersNotifier filtersNotifier, SearchQueryNotifier searchQueryNotifier, AllUsersNotifier? allUsersNotifier) {
+            return allUsersNotifier ?? AllUsersNotifier(filtersNotifier, searchQueryNotifier);
+          },
+        ),
+      ],
       child: const MaterialApp(
         initialRoute: HomePage.routeName,
         onGenerateRoute: generateRoute,
