+++ lib/ui/user_tiles.dart
@@ -1,5 +1,4 @@
-import 'package:first_project/models/user.dart';
-import 'package:first_project/providers/user_notifier.dart';
+import 'package:first_project/notifiers/all_users_notifier.dart';
 import 'package:first_project/ui/user_container.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
@@ -11,7 +10,8 @@ class UserTiles extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    List<User> users = context.watch<UserNotifier>().users;
+    final users = context.watch<AllUsersNotifier>().filterUsers();
+
     return Expanded(
       child: ListView.builder(
         itemCount: users.length,
